# change submodules to use `im::HashMap` instead of `im::OrdMap`