## Description

The [`im` docs](https://docs.rs/im/latest/im/) state that the performance of `HashMap` is similar to that found in `std` but the `ImOrd` is between 2-3x slower than `BTreeMap`. Changing submodules from using `im::OrdMap` to using `im::HashMap` is giving us a nice performance win. 

I tried compiling the libraries project from fluid protocol as a test, it has 19 seperate libraries. This change gives us a 21.25% speed increase. 

Before: Elapsed time: `27.968193458s`
After:  Elapsed time: `22.02603425s`

Running the compile benchmark in LSP is showing a 29% improvement. 

Before: Elapsed time: `3.8328s`
After: Elapsed time: `2.7086s`